### 2025-05-27 a2ddbec
updated references and readme
