+++ paper/paper.bib
@@ -66,14 +66,15 @@
 }
 
 @book{Sullivan2015,
-  title={Introduction to Uncertainty Quantification},
-  author={Sullivan, T.J.},
-  isbn={9783319233956},
-  lccn={2015958897},
-  series={Texts in Applied Mathematics},
-  url={https://books.google.it/books?id=Sik3CwAAQBAJ},
-  year={2015},
-  publisher={Springer International Publishing}
+  title = {Introduction to Uncertainty Quantification},
+  ISBN = {9783319233956},
+  ISSN = {2196-9949},
+  url = {http://dx.doi.org/10.1007/978-3-319-23395-6},
+  DOI = {10.1007/978-3-319-23395-6},
+  journal = {Texts in Applied Mathematics},
+  publisher = {Springer International Publishing},
+  author = {Sullivan,  T.J.},
+  year = {2015}
 }
 
 @book{LeMaitre2010,
@@ -226,7 +227,8 @@
  pages = {105825},
  year = {2023},
  url = {https://doi.org/10.1016/j.envsoft.2023.105825},
- author = {J.D. Jakeman}
+ author = {J.D. Jakeman},
+ doi={10.1016/j.envsoft.2023.105825},
 }
 
 @url{Dakota,
@@ -315,13 +317,28 @@
     URL = {http://www5.in.tum.de/pub/pflueger10spatially.pdf}
 }
 
-@InCollection{DebusschereUQTk:2017,
-  author    = {B. Debusschere and K. Sargsyan and C. Safta and K. Chowdhary},
-  title     = {The Uncertainty Quantification Toolkit (UQTk)},
+@inbook{Debusschere2015,
+  title = {Uncertainty Quantification Toolkit (UQTk)},
+  ISBN = {9783319112596},
+  url = {http://dx.doi.org/10.1007/978-3-319-11259-6_56-1},
+  DOI = {10.1007/978-3-319-11259-6_56-1},
   booktitle = {Handbook of Uncertainty Quantification},
-  editor    = {R. Ghanem and D. Higdon and H. Owhadi},
-  year      = {2017},
-  pages     = {1807--1827},
-  publisher = {Springer},
-  url       = {http://www.springer.com/us/book/9783319123844}
+  publisher = {Springer International Publishing},
+  author = {Debusschere,  Bert and Sargsyan,  Khachik and Safta,  Cosmin and Chowdhary,  Kenny},
+  year = {2015},
+  pages = {1–21}
+}
+
+@article{julia,
+  doi       = {10.1137/141000671},
+  url       = {https://doi.org/10.1137%2F141000671},
+  year      = 2017,
+  month     = {jan},
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title     = {Julia: A Fresh Approach to Numerical Computing},
+  journal   = {{SIAM} Review}
 }
