### 2025-05-28 #8300 @editorialbot
[REVIEW]: SparseGridsKit.jl: Adaptive single- and multi-fidelity sparse grid approximation in Julia
**Submitting author:** @benmkent (<a href="http://orcid.org/0000-0003-4968-7993">Benjamin M. Kent</a>)
**Repository:** https://github.com/benmkent/SparseGridsKit.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.3
