### 2025-06-17 #8300 @dannys4
### For brevity, I will refer to SparseGridsKit.jl as SGK in this review.
Disclaimer: I have my own publicly-available sparse quadrature package that I use internally. It is not published, no one else uses it, and I am not asking for a citation (please do not cite me! It certainly doesn't include any of the impressive interpolation methods included in SGK!), but a few of my suggestions come from my experience developing that codebase. I'll try to highlight where this bias comes in.

A lot of the points I make here are not entirely necessary for me to accept the package for JOSS publication, and I make several of the points because they do not seem addressed in the documentation or current issues. I will try to **bold** the adjustments I would consider absolutely necessary for publication.
# Functionality
- Installation is quite easy, and performing `] test SparseGridsKit` in the REPL works great.
- There are no performance claims.
- A few simple fixes are in order for the tests:
	- when running them locally, **it seems that there are inline function declarations that are defined and re-defined globally that pop up with warnings.**
	- Additionally, **there seem to be domain errors**.
	- I do not believe these affect the results; nevertheless, they should be either namespaced or renamed.
	- Similarly, **it seems like there are some issues with DocTest**---I would imagine that this comes from the fact that SGK is not on the General registry.

Test warnings:
```
WARNING: Method definition p(Any) in module Main at ~/.julia/packages/SparseGridsKit/CbMIO/test/differentiation_test.jl:5 overwritten at ~/.julia/packages/SparseGridsKit/CbMIO/test/differentiation_test.jl:34.
```
```
┌ Warning: Point [-1.2303395598289364, -1.3820028595735336, -1.2673203701438074, -1.245839159380212] is not within domain.
└ @ SparseGridsKit ~/.julia/dev/SparseGridsKit/src/verifyinputs.jl:96
┌ Warning: Point [-1.2303395598289364, -1.3820028595735336, -1.2673203701438074, -1.245839159380212] is not within domain.
└ @ SparseGridsKit ~/.julia/dev/SparseGridsKit/src/verifyinputs.jl:96
```

Doctest warnings:
```
┌ Warning: Unable to determine HTML(edit_link = ...) from remote HEAD branch, defaulting to "master".
│ Calling `git remote` failed with an exception. Set JULIA_DEBUG=Documenter to see the error.
│ Unless this is due to a configuration error, the relevant variable should be set explicitly.
└ @ Documenter ~/.julia/packages/Documenter/2D2ei/src/utilities/utilities.jl:665
```
## Package dependencies:
I am a strong advocate for minimizing necessary dependencies in Julia packages. I often find that package dependencies for corner-case functionality in a Julia package make it difficult to install packages for my application of choice.

- As such, please consider moving the "dependency-heavy" functionality into [package extensions](https://github.com/pebeto/julia_extensions_example), which are promoted for this exact purpose! While package extensions were only introduced in Julia 1.9+, it seems that SGK is only compatible with 1.11+. This would make time-to-first-plot (or time-to-first-grid, perhaps?) significantly faster.
- It seems like, while `Plots` is used in the tests and examples, it is not actually depended upon by the package, which only requires `RecipesBase`. If it is needed for core functionality that I can't see or is needed in the future, consider moving that to a package extension. If not, consider just making a `Project.toml` in each of the `test` and `docs` directories so that SGK is installable without needing `Plots` (I generally prefer Makie anyway).
- On a similar, albeit even more opinionated, note: I would strongly recommend just providing serialized data for the Leja points and making Leja point creation as a package extension. Ninety-nine percent of users will only want the Leja points using either uniform or Chebyshev weighting and will never use more than 100 one-dimensional knots (honestly, 10 is probably enough). Just the size of the Optim package on my machine is 2.2MB (size of `~/.julia/compiled/Optim`, after I've uninstalled every other version of Optim on my system), not even counting any dependencies. Thus, it should be much smaller, not to mention faster, to just include serialized points (and weights?) for the first ~100 knots than it would be to actually find them on the fly.
	- I use serialized points in my package using Julia's `Serial` module, and it's orders of magnitude faster and smaller than additional dependencies. Since `Serial` is in base, and you only work with 1.11+, it is a reasonable choice in your scenario.
- Automatic Differentiation:
	- **It seems that ForwardDiff is not used in your package. Consider removing it from the deps.**
	- Understandably, ReverseDiff is only used to get the derivative of the sparse grid approximations (i.e., [here](https://github.com/benmkent/SparseGridsKit.jl/blob/3ad12a3768edbea07c5a60246fc425013e1c93a7/src/differentiation.jl#L69)). Consider instead using, e.g., [DifferentiationInterface](https://juliadiff.org/DifferentiationInterface.jl/DifferentiationInterface/stable/), which (I believe) is a much lighter dependency and would allow a user to choose the differentiation backend according to their needs (for example, using forward-mode or a faster backend). This also opens up the possibility of using `jacobian` instead of a series of `gradient` calls.
	- More arduous yet, consider just making a package extension for `ChainRules`. I have not done this, nor is this necessary unless you anticipate (or already have) dozens or hundreds of users.
- The `compat` section is quite restrictive. This is not a problem if you use SGK by itself, but would foreseeably be restrictive when used in conjunction with large popular packages or other "bespoke" packages with many dependencies. Perhaps narrow down to the minor versions? At least for the packages above version `1.0`
### Other suggestions (e.g., Quality of life):
- **`truncated_kron` seems to round certain matrix elements below a certain tolerance. While I suspect that this improves numerical stability, it will lose information/signal and provide inaccurate results. This will impact correctness when you translate to and from the spectral sparse grid**. It would behoove you to add an option, e.g., `exact`, to control this (or, perhaps better, setting a package variable to control the tolerance).
- **I would greatly prefer if there was a `verbose` option to relevant methods to turn off the `@info` blocks, which consume my screen when I test the package.**
- It seems that total order multi-index sets (referred to as Smolyak multi-index sets in SGK) are constructed by filtering out a tensor-order multi-index set. This is problematic in the "high-dimensional domains" that SGK claims to approach. For example, in dimension 10 and order 4, consider the difference between the size of the tensor set (1048576) versus the size of the total order set (1001)
	- This is something that comes out of my experience with my own code. There are several ways of implementing it, e.g., recursively or using Kronecker products (my personal choice).
- In some of the code, you use the tilde `~`, which is the bitwise `not`, as well as unpacking into the `~` (as one might do in Matlab). Consider avoiding this to prevent undesirable consequences. Unpacking into tilde redefines it to be that output instead of the bitwise negation.
- A lot of the code could stand to gain from some very cursory profiling. I understand that this package is not intended to be high performing, but the use cases are unfortunately slim given the current state of the package's performance. Since this is not the focus of the package, I will not consider them as prerequisites for publication. I will, however, strongly urge you to spend a day or two profiling the functionality for future researchers interested in using your package.
	- For example, wrapping `gridasptsindices[ii, :]` [here](https://github.com/benmkent/SparseGridsKit.jl/blob/3ad12a3768edbea07c5a60246fc425013e1c93a7/src/sparsegrids.jl#L295) in an `@view` macro, i.e., calling `findfirst(gridasptsindices[ii, :] == uniquerow for uniquerow in eachrow(sparsegridptids))`, can cut the runtime of `sparsegridpoints` dramatically. In the `k=6` example at the bottom of the [Interpolation example](https://benmkent.github.io/SparseGridsKit.jl/stable/sparsegridinterpolation/), it cuts the runtime down from 1.766s to 0.760s on my machine. Generally, the amount of allocations in this code make it fairly restrictive to either low order or low dimension or require unnecessarily large compute resources.
	- In that same example, the time is dominated by calling \=\= on an abstract array, which bolsters Max's argument that creating concrete parametric types (and reducing type inference) would help improve the user experience.
	- `interpolateonsparsegrid` lags tremendously, as noted by Max. The loop [here](https://github.com/benmkent/SparseGridsKit.jl/blob/3ad12a3768edbea07c5a60246fc425013e1c93a7/src/sparsegrids.jl#L417) is a massive part of this lag, and can easily be improved with some loop re-indexing and in-place computations for `feval`, where my revised code is pasted below. This doesn't save a revolutionary amount of time---for my example, it's something like 10%---but I also only thought about it for two minutes and changed <10 lines of code, (seemingly) without breaking any tests.
	- Overall, even relatively sparse grids with few points required hundreds of MB of RAM (or even a few GB) for every one of my experiments, despite the function itself not requiring too much.
```julia
cweightedpolyprod = zeros(length(feval),size(terms,1))
@inbounds for (i, row) = enumerate(eachrow(terms)), k = 1:length(feval)
	targetpt = targetpoints[k]
	cweightedpolyprod[k,i] = cterms[i] * prod(poly[j][row[1][j][1]][row[1][j][2]](targetpt[j]) for j in eachindex(row[1]))
end
@inbounds @fastmath for k in eachindex(feval)
	feval[k] = sum(cweightedpolyprod[k,i] * fongrid[maprowtouniquept[i]] for i in 1:size(terms,1))
end
```

# Documentation
Overall, incredible documentation and examples! I thought all of them were well described and a good introduction to many of the topics in the package. I went through and could reproduce all examples exactly on my own computer!
- In the [Sparse Grid Interpolation](https://benmkent.github.io/SparseGridsKit.jl/stable/sparsegridinterpolation/) example, consider printing the error for the `k=7` example in the same manner as the previous ones (on my computer, I saw an improvement from order 1e-8 to order 1e-10). Really cool though!
	- Nominally, I should note that the Monte Carlo variance on 100 points far exceeds 1e-8, but that's not really the point of this review 😁.
- **Typographical issues:**
	- **There's a weird indentation in some code blocks of a few of the examples (e.g., [Sparse Grid Integration](https://benmkent.github.io/SparseGridsKit.jl/stable/sparsegridintegration/) and [Adaptive Sparse Grids](https://benmkent.github.io/SparseGridsKit.jl/stable/adaptivesparsegrids/)).**
	- **In [Sparse Grid Integration](https://benmkent.github.io/SparseGridsKit.jl/stable/sparsegridintegration/), It looks like there was a slight error typesetting $\sqrt{5}$ instead of $1/\sqrt{5}$.**
	- **In [Multi-fidelity Modelling](https://benmkent.github.io/SparseGridsKit.jl/stable/multifidelity/), it looks like some of the math displays incorrectly.**
## Example Usage
Generally looks good. I tried an example of my own and ran into a small bug (which I reported) but, otherwise, found it rather straightforward to use for simple sparse grid applications. For the adaptive applications, I ran into unexpectedly slow performance when using `LejaPoints`, presumably because there is an inner optimization loop. I also ran into unexpectedly slow performance when interpolating, as suggested in Max's review. I will refer to Max's review and my previous comments on performance and memory allocations for this.

## Contribution guidelines were not included!
**Please include these**.

# Software Paper
My only major critique is the minimal discussion of functional differences between SparseGridsKit.jl and, e.g., `AdaptiveSparsegrids.jl` or `DistributedSparsegrids.jl`. I agree with Max's point above---the quote "offers a more complete set of functionality" is quite vague, and could use more description. While I do not require a full description of what every package does and does not offer, it would be nice to see some specificity on what motivated you to create this tool instead of, e.g., use `AdaptiveSparseGrids.jl` (which has been stable since well before your first commit on SGK).
### Formatting issues:
- Line 21: "Spectral-type polynomials"---could you provide a little more elaboration or description? I believe you are implying that either users often choose to represent the basis using either Lagrange polynomials or the polynomial basis orthogonal w.r.t. the L2 weight of the space. Were there any other representations you were considering?
- Please fix the Dakota citation (using, e.g., [this](https://dakota.sandia.gov/citing-dakota/)).
- Any time there are multiple citations, please collate them (for concision and consistency with journal style)
- Choose either "high dimensional" or "high-dimensional"
