### 2025-07-02 8e644c7
big speed ups to interpolation
