### 2025-07-08 568e1fd
major improvements to interpolation for speed
