+++ src/sparsegrids.jl
@@ -17,7 +17,6 @@ mutable struct SparseGrid
     ptsperlevel
     ptsunique
     sparsegridpts
-    polyperlevel
     domain
 end
 
@@ -40,8 +39,8 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
     dims = sg1.dims
     @assert dims == sg2.dims
 
-    sg1maxmi = length(sg1.polyperlevel)
-    sg2maxmi = length(sg2.polyperlevel)
+    sg1maxmi = length(sg1.ptsperlevel)
+    sg2maxmi = length(sg2.ptsperlevel)
 
     if sg1maxmi >= sg2maxmi
         sg = copy(sg1)
@@ -74,7 +73,7 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
 
     (sparsegridptids, sparsegridpts, maptermstosparsegrid) = sparsegridpoints(terms, sg.pintsmap, sg.ptsunique, size(sg.MI, 1))
 
-    sparsegrid = SparseGrid(dims, MI, terms, sparsegridptids, maptermstosparsegrid, cterms, cmi, sg.pintsmap, sg.ptsperlevel, sg.ptsunique, sparsegridpts, sg.polyperlevel, sg.domain)
+    sparsegrid = SparseGrid(dims, MI, terms, sparsegridptids, maptermstosparsegrid, cterms, cmi, sg.pintsmap, sg.ptsperlevel, sg.ptsunique, sparsegridpts, sg.domain)
 
     return sparsegrid
 end
@@ -94,34 +93,38 @@ end
 
 function sparsegridprecompute(maxmi, knots=CCPoints(), rule=Doubling())
     if isa(knots, Points)
-        ndims=1
-        knots = fill(knots, ndims)
+        nparams=1
+        knots = fill(knots, nparams)
     end
     if isa(rule, Level)
-        rule = fill(rule, ndims)
+        rule = fill(rule, nparams)
     end
     productintegrals = Vector{Any}(undef, length(knots))
     for ii = eachindex(knots)
-        (ptsunique, ptsperlevel, polyperlevel, pintsmap) = sparsegridonedimdata(maxmi, knots[ii], rule[ii], knots[ii].domain)
-        productintegrals[ii] = sparsegridproductintegrals(polyperlevel, maxmi, knots[ii], rule[ii])
+        (ptsunique, ptsperlevel, pintsmap) = sparsegridonedimdata(maxmi, knots[ii], rule[ii], knots[ii].domain)
+        productintegrals[ii] = sparsegridproductintegrals(ptsperlevel, maxmi, knots[ii], rule[ii])
     end
     @debug "Computed weighted L^2 product integrals in each dimension"
     return (; productintegrals)
 end
 
-function sparsegridproductintegrals(polyperlevel, maxmi, knots, rule)
+function sparsegridproductintegrals(ptsperlevel, maxmi, knots, rule)
     @debug "Computing product integrals"
     # Allocate arrays
-    productintegrals = zeros(Float64, maxmi, maxmi, length(polyperlevel[end]), length(polyperlevel[end]))
+    productintegrals = zeros(Float64, maxmi, maxmi, length(ptsperlevel[end]), length(ptsperlevel[end]))
 
     #x, w = gausslegendre(2 * rule(maxmi))
     x,w = knots(2*rule(maxmi))
     wp1x = similar(x)
 
-    pevalx = zeros(Float64, maxmi, length(polyperlevel[end]), length(x))
+    pevalx = zeros(Float64, maxmi, length(ptsperlevel[end]), length(x))
     for ii = 1:maxmi
[…]
