### 2025-07-12 cbac3aa
CompatHelper: bump compat for ForwardDiff to 1, (keep existing compat)
