+++ Project.toml
@@ -22,7 +22,7 @@ Aqua = "0.8.11"
 Documenter = "1.8.0"
 Downloads = "1.6.0"
 FastGaussQuadrature = "1.0.2"
-ForwardDiff = "0.10.38"
+ForwardDiff = "0.10.38, 1"
 LaTeXStrings = "1.4.0"
 LinearAlgebra = "1.11.0"
 MAT = "0.10.7"
