### 2025-07-15 c190d29
Add Leja quadrature points generation and integrate into sparse grid structures
- Implemented `lejapoints` function to use precomputed  Leja quadrature points and weights.
- Updated `LejaPoints` struct to support new type `:precomputed` for using the new Leja points generation method.
- Refactored sparse grid data structures to improve clarity and maintainability, including changes to how terms and unique points are managed.
- Enhanced `SparseGrid` and related functions to utilize the new data structures and improve performance.
- Updated tests for differentiation and sparse grid creation to e
[…]
