+++ src/sparsegrids.jl
@@ -5,19 +5,23 @@ using Polynomials
 import Base.:+
 import Base.:-
 
+mutable struct sparse_grid_data
+    terms::Vector{Any} # Terms in the sparse grid
+    points_to_unique_indices::Matrix{Int} # Mapping from terms to unique points
+    terms_to_grid_points::Vector{Int} # Mapping from terms to grid points
+    coeff_per_term::Vector{Int} # Combination coefficient per term
+    terms_to_unique_points_per_dimension::Vector{Dict{Vector{Int},Int}} # Mapping from terms to unique points per dimension
+    point_sequences_per_dimension::Vector{Vector{Vector{Float64}}}
+    unique_points_per_dimension::Vector{Vector{Float64}} # Unique points per dimension
+end
+
 mutable struct SparseGrid
-    dims # Number of input dimensions
-    MI  # Multi index set
-    terms #
-    sparsegridptids
-    maptermstosparsegrid
-    cterms
-    cmi
-    pintsmap
-    ptsperlevel
-    ptsunique
-    sparsegridpts
-    domain
+    dims::Int # Number of dimensions
+    domain::Vector{Vector{Real}} # Grid domain
+    multi_index_set::Matrix{Int} # Matrix of multi-index columns [β_1 β_2 ...]
+    combination_coeff::Vector{Int} # Combination technique coefficients
+    grid_points::Matrix{Real} # Sparse grid
+    data::sparse_grid_data # Structure for underlying construction data
 end
 
 mutable struct SparseGridApproximation
@@ -39,8 +43,8 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
     dims = sg1.dims
     @assert dims == sg2.dims
 
-    sg1maxmi = length(sg1.ptsperlevel)
-    sg2maxmi = length(sg2.ptsperlevel)
+    sg1maxmi = length(sg1.data.point_sequences_per_dimension)
+    sg2maxmi = length(sg2.data.point_sequences_per_dimension)
 
     if sg1maxmi >= sg2maxmi
         sg = copy(sg1)
@@ -52,36 +56,37 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
         sgb = copy(sg1)
     end
 
-    for (imi, mi) in enumerate(eachcol(sgb.MI))
-        index = findfirst(mi == mi2 for mi2 in eachcol(sga.MI))
+    for (imi, mi) in enumerate(eachcol(sgb.multi_index_set))
+        index = findfirst(mi == mi2 for mi2 in eachcol(sga.multi_index_set))
         if index !== nothing
-            sg.cmi[index] = sg.cmi[index] + sgb.cmi[imi]
+            sg.combination_coeff[index] = sg.combination_coeff[index] + sgb.combination_coeff[imi]
         else
-            push!(sg.cmi, sgb.cmi[imi])
-            sg.MI = hcat(sg.MI, mi)
+            push!(sg.combination_coeff, sgb.combination_coeff[imi])
+            sg.multi_index_set = hcat(sg.multi_index_set, mi)
         end
     end
 
     # Clean zero c points
-    #zeroindices = sg.cmi .!= 0
-    #cmi = sg.cmi[zeroindices]
-    #MI = sg.MI[:, zeroindices]
+    #zeroindices = sg.combination_coeff .!= 0
+    #combination_coeff = sg.combination_coeff[zeroindices]
+    #multi_index_set = sg.multi_index_set[:, zeroindices]
 
-    (MI,cmi) = sortmi(sg.MI, sg.cmi)
+    (multi_index_set,combination_coeff) = sortmi(sg.multi_index_set, sg.combination_coeff)
 
-    (terms, cterms, cmi, MI) = sparsegridterms(MI, sg.ptsperlevel, cmi=sg.cmi)
[…]
