### 2025-07-22 603891d
Refactor sparse grid point mapping and improve computesparsegridc function
