### 2025-07-23 ef1e4c2
Refactor integration functions to remove precomputed Lagrange integrals and add quadrature weight computation
