+++ test/multifidelity_tests.jl
@@ -23,13 +23,11 @@ using SparseGridsKit
     f_true = [10.0^-z[1] + 100.0*z[2]+10.0*z[3]+z[4] for z in get_grid_points(sg)]
     @test f_eval ≈ f_true
 
-    pcl = precompute_lagrange_integrals(5, knots, rule)
-
-    integral_result = integrate_on_sparsegrid(sg, f_eval, pcl)
+    integral_result = integrate_on_sparsegrid(sg, f_eval)
     @test integral_result ≈ 0.001
 
     (sg, f_on_Z) = adaptive_sparsegrid(f_wrapped, nfid+nparams; maxpts = 100, proftol=1e-4, rule = rule, knots = knots, θ=1e-4, type=:deltaint, costfunction=α->10.0^prod(α))
 
-    integral_result = integrate_on_sparsegrid(sg, f_on_Z, pcl)
+    integral_result = integrate_on_sparsegrid(sg, f_on_Z)
     @test abs(integral_result) ≤ 1e-4 
 end
