### 2025-07-29 1363734
Update Dakota documentation and refine UQTk entry in bibliography; enhance quadrature weight computation in SparseGrid
