+++ paper/paper.bib
@@ -234,7 +234,7 @@
 @url{Dakota,
   author ={B. M. Adams and W. J. Bohnhoff and K. R. Dalbey and M. S. Ebeida and J. P. Eddy and M. S. Eldred and R. W. Hooper and P. D. Hough and K. T. Hu and J. D. Jakeman and M. Khalil and K. A. Maupin and J. A. Monschke and E. E. Prudencio and E. M. Ridgway and P. Robbe and A. A. Rushdi and D. T. Seidl and J. A. Stephens and L. P. Swiler and J. G. Winokur},
   title = {Dakota 6.21.0 documentation. Technical Report SAND2024-15492O.},
-  publisher = {Sandia National Laboratories, Albuquerque, NM},
+  publisher = {Sandia National Laboratories, Albuquerque, NM, November 2024},
   year = {2024},
   month = nov,
   url = {http://snl-dakota.github.io}
@@ -317,16 +317,28 @@
     URL = {http://www5.in.tum.de/pub/pflueger10spatially.pdf}
 }
 
-@inbook{Debusschere2015,
+% @inbook{Debusschere2015,
+%   title = {Uncertainty Quantification Toolkit (UQTk)},
+%   ISBN = {9783319112596},
+%   url = {http://dx.doi.org/10.1007/978-3-319-11259-6_56-1},
+%   DOI = {10.1007/978-3-319-11259-6_56-1},
+%   booktitle = {Handbook of Uncertainty Quantification},
+%   publisher = {Springer International Publishing},
+%   author = {Debusschere,  Bert and Sargsyan,  Khachik and Safta,  Cosmin and Chowdhary,  Kenny},
+%   year = {2015},
+%   pages = {1–21}
+% }
+
+@inbook{DebusschereUQTk:2017,
   title = {Uncertainty Quantification Toolkit (UQTk)},
-  ISBN = {9783319112596},
-  url = {http://dx.doi.org/10.1007/978-3-319-11259-6_56-1},
-  DOI = {10.1007/978-3-319-11259-6_56-1},
+  ISBN = {9783319123851},
+  url = {http://dx.doi.org/10.1007/978-3-319-12385-1_56},
+  DOI = {10.1007/978-3-319-12385-1_56},
   booktitle = {Handbook of Uncertainty Quantification},
   publisher = {Springer International Publishing},
   author = {Debusschere,  Bert and Sargsyan,  Khachik and Safta,  Cosmin and Chowdhary,  Kenny},
-  year = {2015},
-  pages = {1–21}
+  year = {2017},
+  pages = {1807–1827}
 }
 
 @article{julia,
