### 2025-07-30 ebf6fec
Refactor adaptive sparse grid functions and improve quadrature weight computation
