+++ simpletest.jl
@@ -23,9 +23,8 @@ function sparse_grid(N::Int,pointprops; nlevel=6,RT=Float64,CT=Float64)
 	return asg
 end
 
-(sg, f_on_Z) = adaptive_sparsegrid(fun1, nparams, maxpts=10_000, proftol=1e-3)
-pcl = precompute_lagrange_integrals(7)
-integral_result = integrate_on_sparsegrid(sg, f_on_Z, pcl)
+(sg, f_on_Z) = adaptive_sparsegrid(fun1, nparams, maxpts=10_000, proftol=1e-5)
+integral_result = integrate_on_sparsegrid(sg, f_on_Z)
 
 asg = sparse_grid(2, @SVector [1,1])
 init_weights!(asg, fun1)
