+++ src/sparsegrids.jl
@@ -6,13 +6,14 @@ import Base.:+
 import Base.:-
 
 mutable struct sparse_grid_data
-    terms::Vector{Any} # Terms in the sparse grid
+    terms::Vector{Tuple}
     points_to_unique_indices::Matrix{Int} # Mapping from terms to unique points
     terms_to_grid_points::Vector{Int} # Mapping from terms to grid points
     coeff_per_term::Vector{Int} # Combination coefficient per term
     terms_to_unique_points_per_dimension::Vector{Dict{Vector{Int},Int}} # Mapping from terms to unique points per dimension
     point_sequences_per_dimension::Vector{Vector{Vector{Float64}}}
     unique_points_per_dimension::Vector{Vector{Float64}} # Unique points per dimension
+    weight_sequences_per_dimension::Vector{Vector{Vector{Float64}}}
 end
 
 mutable struct SparseGrid
@@ -29,6 +30,7 @@ end
 
 function SparseGrid(dims, domain, multi_index_set, combination_coeff, grid_points, knots, rules, data)
     sg = SparseGrid(dims, domain, multi_index_set, combination_coeff, grid_points, [], knots, rules, data)
+    compute_quadrature_weights!(sg)
     return sg
 end
 
@@ -85,7 +87,7 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
 
     (points_to_unique_indices, grid_points, terms_to_grid_points) = sparsegridpoints(terms, sg.data.terms_to_unique_points_per_dimension, sg.data.unique_points_per_dimension, size(sg.multi_index_set, 1))
 
-    data = sparse_grid_data(terms, points_to_unique_indices, terms_to_grid_points, coeff_per_term, terms_to_unique_points_per_dimension, sg.data.point_sequences_per_dimension, sg.data.unique_points_per_dimension)
+    data = sparse_grid_data(terms, points_to_unique_indices, terms_to_grid_points, coeff_per_term, terms_to_unique_points_per_dimension, sg.data.point_sequences_per_dimension, sg.data.unique_points_per_dimension, weight_sequences_per_dimension)
 
     sparsegrid = SparseGrid(dims, sg.domain, multi_index_set, combination_coeff, grid_points, sg.knots, sg.rules, data)
     return sparsegrid
@@ -114,7 +116,7 @@ function sparsegridprecompute(maxmi, knots=CCPoints(), rule=Doubling())
     end
     productintegrals = Vector{Any}(undef, length(knots))
     for ii = eachindex(knots)
-        (unique_points_per_dimension, point_sequences_per_dimension, terms_to_unique_points_per_dimension) = sparsegridonedimdata(maxmi, knots[ii], rule[ii], knots[ii].domain)
+        (unique_points_per_dimension, point_sequences_per_dimension, terms_to_unique_points_per_dimension, weight_sequences_per_dimension) = sparsegridonedimdata(maxmi, knots[ii], rule[ii], knots[ii].domain)
         productintegrals[ii] = sparsegridproductintegrals(point_sequences_per_dimension, maxmi, knots[ii], rule[ii])
     end
     @debug "Computed weighted L^2 product integrals in each dimension"
@@ -268,11 +270,12 @@ function createsparsegrid(multi_index_set; rule=Doubling(), knots=CCPoints())
     end
     unique_points_per_dimension = Vector{Vector{Float64}}(undef,dims)
     point_sequences_per_dimension = Vector{Vector{Vector{Float64}}}(undef,dims)
+    weight_sequences_per_dimension = Vector{Vector{Vector{Float64}}}(undef,dims)
     terms_to_unique_points_per_dimension = Vector{Dict{Vector{Int},Int}}(undef,dims)
[…]
