+++ .github/workflows/test.yaml
@@ -2,7 +2,6 @@ name: Test
 
 on:
   push:
-    tags: '*'
     paths:
       - 'src/**'
       - 'test/**'
