### 2025-08-05 0ae5c80
all tests pass again, optimised interpolation
