### 2025-08-06 add78bf
Massively optimised code to reduce allocations, and number of lagrange polynomial evaluations
Refactor code for clarity and performance improvements; update multi-index handling and fix logical errors in input verification.
