+++ src/sparsegrids.jl
@@ -5,13 +5,17 @@ using Polynomials
 import Base.:+
 import Base.:-
 
-mutable struct SparseTerm
-    data::Vector{SVector{2, Int}}
+# mutable struct SparseTerm
+#     data::Vector{SVector{2, Int}}
+# end
+
+struct SparseTerm{N}
+    data::NTuple{N, SVector{2,Int}}
 end
 
-mutable struct sparse_grid_data
-    terms::Vector{SparseTerm}
-    points_to_unique_indices::Matrix{Int} # Mapping from terms to unique points
+mutable struct sparse_grid_data{N}
+    terms::Vector{SparseTerm{N}}
+    points_to_unique_indices::Vector{NTuple{N,Int}} # Mapping from terms to unique points
     terms_to_grid_points::Vector{Int} # Mapping from terms to grid points
     coeff_per_term::Vector{Int} # Combination coefficient per term
     terms_to_unique_points_per_dimension::Vector{Dict{Vector{Int},Int}} # Mapping from terms to unique points per dimension
@@ -22,13 +26,13 @@ end
 
 mutable struct SparseGrid
     dims::Int # Number of dimensions
-    domain::Vector{Vector{Real}} # Grid domain
+    domain::Vector{Vector{Float64}} # Grid domain
     multi_index_set::Matrix{Int} # Matrix of multi-index columns [β_1 β_2 ...]
     combination_coeff::Vector{Int} # Combination technique coefficients
     grid_points::Matrix{Real} # Sparse grid
     quadrature_weights::Vector{Float64} # Quadrature weights
-    knots::Vector{Any}
-    rules::Vector{Any}
+    knots::Vector{Points}
+    rules::Vector{Level}
     data::sparse_grid_data # Structure for underlying construction data
 end
 
@@ -91,7 +95,7 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
 
     (points_to_unique_indices, grid_points, terms_to_grid_points) = sparsegridpoints(terms, sg.data.terms_to_unique_points_per_dimension, sg.data.unique_points_per_dimension, size(sg.multi_index_set, 1))
 
-    data = sparse_grid_data(terms, points_to_unique_indices, terms_to_grid_points, coeff_per_term, terms_to_unique_points_per_dimension, sg.data.point_sequences_per_dimension, sg.data.unique_points_per_dimension, weight_sequences_per_dimension)
+    data = sparse_grid_data{dims}(terms, points_to_unique_indices, terms_to_grid_points, coeff_per_term, terms_to_unique_points_per_dimension, sg.data.point_sequences_per_dimension, sg.data.unique_points_per_dimension, weight_sequences_per_dimension)
 
     sparsegrid = SparseGrid(dims, sg.domain, multi_index_set, combination_coeff, grid_points, sg.knots, sg.rules, data)
     return sparsegrid
@@ -293,7 +297,7 @@ function createsparsegrid(multi_index_set; rule=Doubling(), knots=CCPoints())
     (points_to_unique_indices, grid_points, terms_to_grid_points) = sparsegridpoints(grid, terms_to_unique_points_per_dimension, unique_points_per_dimension, dims)
     @debug "Created sparse grid term to sparse grid points mappings"
 
-    data = sparse_grid_data(grid, points_to_unique_indices, terms_to_grid_points, coeff_per_term, terms_to_unique_points_per_dimension, point_sequences_per_dimension, unique_points_per_dimension, weight_sequences_per_dimension)
+    data = sparse_grid_data{dims}(grid, points_to_unique_indices, terms_to_grid_points, coeff_per_term, terms_to_unique_points_per_dimension, point_sequences_per_dimension, unique_points_per_dimension, weight_sequences_per_dimension)
 
[…]
