### 2025-08-06 6e0a7ba
fix for non vector objects
