### 2025-08-06 348afb8
fix: correct syntax in differentiation examples and improve error handling in interpolation
Documentation should now generate properly
