### 2025-08-06 2cdfc17
fixes to julia doccumentation and update of project.toml
