+++ src/sparsegridplots.jl
@@ -1,5 +1,4 @@
 using RecipesBase
-using StatsBase
 
 """
     f(p::Points; n = 11)
@@ -16,49 +15,6 @@ Recipe to plot scatter of knots and weights.
     xlabel --> "Knots"
     ylabel --> "Weight"
     x,w = p(n)
-end
-
-"""
-    misetplot(miset::MISet)
-
-    Scatter plot of MI in pairs of dimensions
-
-# Arguments
-- `miset` : Multi-index set to plot
-"""
-@userplot MISetPlot
-@recipe function f(h::MISetPlot)
-    miset = h.args[1]
-    @assert isa(miset, MISet)
-    data = hcat(get_mi(miset)...)'
-
-    n, k = size(data, 1), size(data, 2)
-    
-    title --> "Multi-Index Pair Plot"
-
-    # Set up subplots (creating a k x k grid)
-    layout := @layout (k,k)  # Create a k x k layout grid
-
-    # Loop through the data to create scatter plots
-    for jj = 1:k
-        for ii = 1:jj  # Only plot the lower triangle and diagonal
-            x, y = data[:, ii], data[:, jj]
-            # # Marker size for frequency
-            # points = [(x[i], y[i]) for i in 1:length(x)]
-            # point_counts = countmap(points)
-            # unique_points = keys(point_counts)
-            # frequencies = values(point_counts)
-            # sizes = [100 * frequencies[findfirst(isequal(p), unique_points)] for p in points]
-            @series begin
-                subplot := ii + k*(jj-1)  # Assign subplot to the correct position
-                seriestype := :scatter  # Define the plot type
-                xlabel --> "Variable $ii"
-                ylabel --> "Variable $jj"
-                # Return x,y
-                x,y
-            end
-        end
-    end
 end
 
 """
@@ -80,16 +36,16 @@ end
     xlabel --> "Parameter "*string(targetdims[1])
     x = points_matrix[:,targetdims[1]]
     y = zeros(size(x))
-    z = zeros(size(x))
     if sg.dims > 1
         ylabel --> "Parameter "*string(targetdims[2])
         y = points_matrix[:,targetdims[2]]
+        x,y
     end
     if sg.dims > 2
         zlabel --> "Parameter "*string(targetdims[3])
         z = points_matrix[:,targetdims[3]]
+        x,y,z
     end
-    x,y,z
 end
 
 """
@@ -102,28 +58,36 @@ Recipe for plotting approximations based on sparse grids. For more than two dime
 @recipe function f(sga::Union{SparseGridApproximation,SpectralSparseGridApproximation}; targetdims=[1,2])
     midpoint = mean.(sga.domain)
 
-    subdomain1 = sga.domain[targetdims[1]]
-    subdomain2 = sga.domain[targetdims[2]]
-
-    function  modifiedmidpoint(x,i,y,j)
-        value = copy(midpoint)
-        value[i] = x
-        value[j] = y
-        return value        
-    end
-
     n = 100
-    x =  range(subdomain1[1], stop=subdomain1[2], length=n)
-    y =  range(subdomain2[1], stop=subdomain2[2], length=n)
-    xy = [modifiedmidpoint(xi,targetdims[1],yi,targetdims[2]) for xi in x for yi in y]
+    if length(midpoint) == 1
+        subdomain1 = sga.domain[targetdims[1]]
+        x =  range(subdomain1[1], stop=subdomain1[2], length=n)
+        y = sga.(x)
+        title --> "Sparse Grid Approximation"
+        xlabel --> "Parameter "*string(targetdims[1])
+        x,y
+    else
+        subdomain1 = sga.domain[targetdims[1]]
[…]
