### 2025-08-07 88a7f5c
optimise generating multi-index sets
