### 2025-08-07 154fa1e
bump version to 0.2.0 in Project.toml
there have been changes in API since 0.1
