### 2025-08-07 7063308
updated paper for JOSS
