+++ paper/paper.md
@@ -21,30 +21,47 @@ bibliography: paper.bib
 ---
 
 # Summary
-Approximating functions with high-dimensional domains is crucial for modern scientific and engineering problems.
+Approximating functions with high-dimensional domains is crucial in modern scientific and engineering problems.
 An example of this is constructing surrogate models for quantities of interest in high-dimensional, parametrized PDE problems.
 These surrogate models are constructed to provide computationally inexpensive yet accurate approximations that can be used in applications such as uncertainty quantification, optimization, and parameter estimation [@UQHandbook2017].
-Surrogates can be constructed using global polynomial approximation on the parameter space, and a common approach is the use of sparse grid approximation techniques.
+For suitably smooth functions, accurate surrogates can be constructed using global polynomial approximation techniques over the parameter domain, and a common approach is the use of sparse grid polynomial approximation.
 In particular, sparse grid polynomial interpolation techniques enable practitioners to approximate solutions to parametric problems in a non-intrusive manner using existing numerical solvers.
 
 `SparseGridsKit.jl` provides a Julia toolbox to manually and adaptively construct sparse grid polynomial approximations [@julia].
 Interpolation and quadrature routines allow evaluation and integration of the surrogate models.
-Multi-fidelity approximation via the multi-index stochastic collocation algorithm is also possible [@HajiAli2016] [@Jakeman2019] [@Piazzola2022].
-Approximations can be represented either in a basis of Lagrange interpolation polynomials or in a basis of spectral-type polynomials.
+Multi-fidelity approximation via the multi-index stochastic collocation algorithm is also possible [@HajiAli2016; @Jakeman2019; @Piazzola2022].
+Approximations can be represented either in a basis of global Lagrange interpolation polynomials or in a basis of domain appropriate spectral-type global polynomials (e.g.\ Legendre, Chebyshev, Hermite...).
 
 # Statement of need
-Sparse grid approximation is a well-developed methodology and is featured in many survey articles and textbook chapters, e.g., [@Bungartz2004],[@LeMaitre2010],[@Schwab2011],[@Cohen2015],[@Sullivan2015].
-The need for sparse grid surrogate modelling is demonstrated by its use in many applications, from simpler elliptic and parabolic PDEs to complex practical engineering problems, e.g., \ [@Piazzola2021],[@Piazzola2022],[@Li2024].
+Sparse grid approximation is a well-developed methodology and is featured in many survey articles and textbook chapters, e.g., [@Bungartz2004; @LeMaitre2010; @Schwab2011; @Cohen2015; @Sullivan2015].
+The need for sparse grid surrogate modelling is demonstrated by its use in many applications, from simpler elliptic and parabolic PDEs to complex practical engineering problems, e.g., \ [@Piazzola2021; @Piazzola2022; @Li2024].
 The `SparseGridsKit.jl` implementation offers a rich set of features to enable this.
 
-Specifically, `SparseGridsKit.jl` is a Julia implementation of sparse grid approximation methods.
-This offers
+Specifically, `SparseGridsKit.jl` is a Julia implementation of adaptive sparse grid global polynomial approximation methods.
+Development was motivated by the desire for approximation code that closely resembles the mathematical literature to aid algorithm development and analysis.
 
[…]
