+++ paper/paper.bib
@@ -354,3 +354,41 @@
   title     = {Julia: A Fresh Approach to Numerical Computing},
   journal   = {{SIAM} Review}
 }
+
+@article{Bittens2023,
+  title = {DistributedSparseGrids.jl: A Julia library implementing
+an Adaptive Sparse Grid collocation method},
+  volume = {8},
+  ISSN = {2475-9066},
+  url = {http://dx.doi.org/10.21105/joss.05003},
+  DOI = {10.21105/joss.05003},
+  number = {83},
+  journal = {Journal of Open Source Software},
+  publisher = {The Open Journal},
+  author = {Bittens,  Maximilian and Gates,  Robert L.},
+  year = {2023},
+  month = mar,
+  pages = {5003}
+}
+
+@inbook{Nobile2016,
+  title = {An Adaptive Sparse Grid Algorithm for Elliptic PDEs with Lognormal Diffusion Coefficient},
+  ISBN = {9783319282626},
+  ISSN = {2197-7100},
+  url = {http://dx.doi.org/10.1007/978-3-319-28262-6_8},
+  DOI = {10.1007/978-3-319-28262-6_8},
+  booktitle = {Sparse Grids and Applications - Stuttgart 2014},
+  publisher = {Springer International Publishing},
+  author = {Nobile,  Fabio and Tamellini,  Lorenzo and Tesei,  Francesco and Tempone,  Raúl},
+  year = {2016},
+  pages = {191–220}
+}
+
+@inproceedings{genz,
+  author =  {Genz, Alan C.},
+  title = {Testing multidimensional integration routines},
+  booktile = {Tools, Methods, and Languages for Scientific and Engineering Computation},
+  editor = {B. Ford, J.C. Rault and F. Thomasset},
+  year = 1984,
+  pages = {81–94},
+}
