### 2025-08-07 cfc614b
remove simpletest.jl
