### 2025-08-07 #8300 @benmkent
Thank you very much for the feedback @fvdugo. I have hopefully addressed your comments, please see below.

- Functionality: As suggested, the package is in the process of being added to the Julia registry. I will update the GitHub repository and documentation once the registration is merged.
- Statement of need: I have clarified this more explicitly in the paper. The motivation was to create a codebase that stays close to the mathematical literature, enabling algorithm development and analysis for global polynomial approximation using sparse grids.
- State of the field: I have now emphasized the main differences and motivations behind SparseGridsKit.jl compared to existing software. Most notably, this includes the use of global polynomial approximation and the structure of the adaptive algorithm, which is designed to facilitate algorithm development.
- Quality of writing: I have added an example to the paper demonstrating adaptive sparse grid approximation, interpolation, and integration. I’d be happy to include additional examples if you think the paper would benefit further.
