### 2025-08-07 #8300 @benmkent
@baxmittens thank you very much for your detailed review of both the paper and the code.

I will first address your remarks and example.
Thank you very much for exposing the inefficiency in my code - in the test problems I had previously run it had not been obvious to me.

Both `adaptive_sparsegrid` and `interpolate_on_sparsegrid` have now been significantly improved in terms of runtime and memory usage. While they are still slower than your implementation, performance is now more competitive.
In particular, the interpolation may remain slower due to the use of global polynomial basis functions, in contrast to local basis functions.

To compare, my code now adaptively generates 5377 points via
```
@time (sg, f_on_Z) = adaptive_sparsegrid(fun1, nparams, maxpts=10_000, proftol=1e-5)
  2.563692 seconds (38.64 M allocations: 1.819 GiB, 19.63% gc time)
```
For your code with 5165 points
```
@time begin
       asg = sparse_grid(2, @SVector [1,1])
       init_weights!(asg, fun1)
       for i = 1:20
               cpts = generate_next_level!(asg, 1e-3, 20)
               init_weights!(asg, collect(cpts), fun1)
       end
       end
  0.108811 seconds (468.43 k allocations: 21.878 MiB)
```
Most of the additional cost in my implementation is from repeatedly constructing SparseGrid structures during error estimation. I do plan to optimize this, though current performance is acceptable for my applications.

With comparable grid sizes, interpolation now performs as follows:
```
julia> @time zs1 = interpolate_on_sparsegrid(sg, f_on_Z, zvals[:]);

 14.063538 seconds (116.83 k allocations: 326.570 MiB, 2.47% gc time)

julia> @time zs2 = [interpolate(asg,[x,y]) for x in xs, y in xs];
  0.441403 seconds (2.00 M allocations: 77.726 MiB, 24.33% compilation time)
```
The majority of the time is spent in evaluating Lagrange interpolation polynomials.
From profiling the allocations come from initializing the output vector - this possibly can be optimised further but it is not immediately obvious to me.

##
The `SparseGrid` structure is not typed as it does not hold the function evaluations.
The function evaluations can take any type, and for objects such as Vectors inplace operations now minimize the required allocations (e.g. running your provided example should be reasonably efficient, if not as fast as your code).
Note that in the documentation there is an example where the evaluations are `Fun` objects from the `ApproxFun.jl` package.
It is possible to interpolate this, however it is much more expensive as allocations occur even for `1.0*f` for `f::Fun`.

It is possible that I should `type` the `SparseGridApproximation` structure. I will consider this in the future.

##
In the paper, I have hopefully made a clearer comparison to your code (and other related packages).
In particular, I highlight:

- The use of global polynomial basis functions in my code;
- The design goal of matching the mathematical structure of adaptive algorithms, which we found useful for developing new algorithmic variants.

This was the key motivation for our work, as we have been developing related algorithmic enhancements.
