+++ src/sparsegrids.jl
@@ -5,14 +5,21 @@ using Polynomials
 import Base.:+
 import Base.:-
 
-# mutable struct SparseTerm
-#     data::Vector{SVector{2, Int}}
-# end
+"""
+    SparseTerm{N}
 
+Structure representing a sparse grid `term` with `N` dimensions.
+
+"""
 struct SparseTerm{N}
-    data::NTuple{N, SVector{2,Int}}
+    data::NTuple{N, SVector{2,Int}} # Tuple containing `N` static vectors of two integers each.
 end
 
+"""
+    sparse_grid_data{N}
+
+Underlying data for a sparse grid of dimension `N`.
+"""
 mutable struct sparse_grid_data{N}
     terms::Vector{SparseTerm{N}}
     points_to_unique_indices::Vector{NTuple{N,Int}} # Mapping from terms to unique points
@@ -24,6 +31,11 @@ mutable struct sparse_grid_data{N}
     weight_sequences_per_dimension::Vector{Vector{Vector{Float64}}}
 end
 
+"""
+    SparseGrid
+
+Sparse grid structure.
+"""
 mutable struct SparseGrid
     dims::Int # Number of dimensions
     domain::Vector{Vector{Float64}} # Grid domain
@@ -36,27 +48,52 @@ mutable struct SparseGrid
     data::sparse_grid_data # Structure for underlying construction data
 end
 
+"""
+    SparseGrid(dims, domain, multi_index_set, combination_coeff, grid_points, knots, rules, data)
+
+Constructs a `SparseGrid` object with quadrature weights.
+"""
 function SparseGrid(dims, domain, multi_index_set, combination_coeff, grid_points, knots, rules, data)
     sg = SparseGrid(dims, domain, multi_index_set, combination_coeff, grid_points, [], knots, rules, data)
     compute_quadrature_weights!(sg)
     return sg
 end
 
+"""
+    SparseGridApproximation
+
+Function approximation consisting of sparse grid and function values on the grid.
+"""
 mutable struct SparseGridApproximation
     sg::SparseGrid
     fongrid::Vector
     domain::Vector
 end
 
+"""
+    SparseGridApproximation
+
+Constructs a SparseGridApproximation.
+"""
 function SparseGridApproximation(sg, fongrid)
     SparseGridApproximation(sg,fongrid,sg.domain)
 end
 
+"""
+    (sga::SparseGridApproximation)(x)
+
+Evaluate SparseGridApproximation at point `x`.
+"""
 function (sga::SparseGridApproximation)(x)
     @assert length(x) == sga.sg.dims
     return interpolate_on_sparsegrid(sga.sg, sga.fongrid, [x])[1]
 end
 
+"""
+    Base.:+(sg1::SparseGrid, sg2::SparseGrid)
+
+Add two `SparseGrid` objects by adding their common combination coefficients and merging their multi-index sets.
+"""
 function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
     dims = sg1.dims
     @assert dims == sg2.dims
@@ -84,11 +121,6 @@ function Base.:+(sg1::SparseGrid, sg2::SparseGrid)
         end
     end
 
-    # Clean zero c points
-    #zeroindices = sg.combination_coeff .!= 0
-    #combination_coeff = sg.combination_coeff[zeroindices]
-    #multi_index_set = sg.multi_index_set[:, zeroindices]
-
     (multi_index_set,combination_coeff) = sortmi(sg.multi_index_set, sg.combination_coeff)
 
[…]
