+++ .github/workflows/DocumentationUserRedeploy.yml
@@ -1,0 +1,30 @@
+name: Documentation (User Triggered Full Redeploy)
+
+on:
+  workflow_dispatch:
+
+jobs:
+  docs:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0   # Need full history + all tags
+      - uses: julia-actions/setup-julia@v2
+        with:
+          version: '1.10'  # or your preferred Julia
+      - name: Install dependencies
+        run: |
+          julia --project=docs -e '
+            using Pkg;
+            Pkg.develop(PackageSpec(path=pwd()));
+            Pkg.instantiate();
+          '
+      - name: Build & deploy all versions
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          julia --project=docs -e '
+            using Documenter;
+            include("docs/make.jl");
+          '
