### 2025-08-11 d7ed643
Update Project.toml
Fix compatability requirements
