+++ Project.toml
@@ -31,8 +31,8 @@ Plots = "^1"
 PolyChaos = "0.2"
 Polynomials = "^4"
 RecipesBase = "^1"
-Serialization = "^1.11"
-SparseArrays = "^1.11"
+Serialization = "^1.0"
+SparseArrays = "^1.9"
 StaticArrays = "^1"
 Test = "^1"
 julia = "^1.9"
