+++ Project.toml
@@ -19,7 +19,7 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 [compat]
 ApproxFun = "^0.13"
 Aqua = "^0.8"
-DifferentiationInterface = "^0.7"
+DifferentiationInterface = "^0.6"
 Documenter = "^1"
 Downloads = "^1"
 FastGaussQuadrature = "^1"
