### 2025-08-11 368bd92
simplify make.jl hopefully to fix documentation
