+++ v0.2.0-docs/siteinfo.js
@@ -1,0 +1,2 @@
+var DOCUMENTER_CURRENT_VERSION = "v0.2.0-docs";
+var DOCUMENTER_IS_DEV_VERSION = false;
