+++ v0.2.0-docs/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"spectralsparsegrids/#Spectral-Sparse-Grids","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"","category":"section"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"The sparse grid approximations are polynomial approximations, with a basis of Lagrange interpolation polynomials. The approximation can be written in any suitable polynomial basis, for example we could use the Chebyshev polynomials.","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"The change of basis is implemented exploiting the ApproxFun.jl package. The underlying interpolation polynomials used in SparseGridsKit.jl are constructed using Funs. To change basis we use an ApproxFun.jl expansion into coefficients and form a tensor product. To promote sparsity this is done using a truncated Kronecker product. The polynomial coefficients are assembled with corresponding polynomial degrees and spaces in a callable SpectralSparseGridApproximation.","category":"page"},{"location":"spectralsparsegrids/#Examples","page":"Spectral Sparse Grids","title":"Examples","text":"","category":"section"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"An adaptive sparse grid approximation for a Genz test function is created. ","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"using SparseGridsKit\nn = 8\nC = 1.0\nW = 0.0\nT = \"exponentialdecay\"\nN = \"gaussianpeak\"\nf = genz(n::Int, C::Float64, W::Float64, T::String, N::String)\n# Approximate\n(sg, f_on_Z) = adaptive_sparsegrid(f, n)\nf_sg = SparseGridApproximation(sg,f_on_Z)","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"The approximation f_sg is a representation formed as a linear combination of Lagrange interpolation polynomials. This is converted to a spectral representation using convert_to_spectral_approximation.","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"f_spectral = convert_to_spectral_approximation(sg, f_on_Z)","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"This still has an n dimensional domain.","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"f_spectral.dims","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"The representation of the tensor of polynomials is compactly stored as a SparseVector. To index into this we require the storage dimension for each domain dimension.","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"f_spectral.expansiondimensions","category":"page"},{"location":"spectralsparsegrids/","page":"Spectral Sparse Grids","title":"Spectral Sparse Grids","text":"This demonstrates that more polynomial
[…]
