+++ v0.2.1/refs.bib
@@ -1,0 +1,86 @@
+@article{Barthelmann2000,
+	title        = {High dimensional polynomial interpolation on sparse grids},
+	author       = {Volker Barthelmann and Erich Novak and Klaus Ritter},
+	year         = 2000,
+	journal      = {Advances in Computational Mathematics},
+	publisher    = {Springer Science and Business Media {LLC}},
+	volume       = 12,
+	number       = 4,
+	pages        = {273--288},
+	doi          = {10.1023/a:1018977404843},
+}
+@article{Clenshaw1960,
+	title        = {A method for numerical integration on an automatic computer},
+	author       = {Clenshaw, C. W. and Curtis, A. R.},
+	year         = 1960,
+	month        = dec,
+	journal      = {Numerische Mathematik},
+	publisher    = {Springer Science and Business Media {\{}LLC{\}}},
+	volume       = 2,
+	number       = 1,
+	pages        = {197--205},
+	doi          = {10.1007/BF01386223},
+	issn         = {0029599X},
+	abstract     = {A new method for the numerical integration of a "well-behaved" function over a finite range of argument is described. It consists essentially of expanding the integrand in a series of Chebyshev polynomials, and integrating this series term by term. Illustrative examples are given, and the method is compared with the most commonly-used alternatives, namely Simpson's rule and the method of Gauss. {\textcopyright} 1960 Springer-Verlag.},
+}
+@article{Gerstner1998,
+	title        = {Numerical integration using sparse grids},
+	author       = {Gerstner, Thomas and Griebel, Michael},
+	year         = 1998,
+	journal      = {Numerical algorithms},
+	volume       = 18,
+	number       = 3,
+	pages        = {209--232},
+	doi          = {10.1023/A:1019129717644},
+	issn         = {1017-1398},
+	abstract     = {We present new and review existing algorithms for the numerical integration of multi-variate functions defined over d-dimensional cubes using several variants of the sparse grid method first introduced by Smolyak [49]. In this approach, multivariate quadrature formulas are constructed using combinations of tensor products of suitable one-dimensional formulas. The computing cost is almost independent of the dimension of the problem if the function under consideration has bounded mixed derivatives. We suggest the usage of extended Gauss (Patterson) quadrature formulas as the one-dimensional basis of the construction and show their superiority in comparison to previously used sparse grid approaches based on the trape-zoidal, Clenshaw–Curtis and Gauss rules in several numerical experiments and applications. For the computation of path integrals further improvements can be obtained by combining generalized Smolyak quadrature with the Brownian bridge construction.},
+	file         = {:home/ben/Documents/GitHub/phd/Papers/Gerstner1998/Gerstner-Griebel1998{\_}Article{\_}NumericalIntegrationUsingSpars.pdf:pdf},
+	keywords     = {Multivariate numerical quadrature,Smolyaks construction,complexity,curse of dimension,d-dimensional cube,dimensional formula,gauss rule,mixed derivative,multivariate function,multivariate quadrature formula,numerical integration,one-dimensional basis,quadrature formula,review algorithm,several numerical experiment,several variant,sparse grid,sparse grid approach,sparse grid method,sparse grids,tensor product}
+}
+@article{Gerstner2003,
+	title        = {Dimension--Adaptive Tensor--Product Quadrature},
+	author       = {Gerstner, T. and Griebel, M.},
+	year         = 2003,
+	month        = aug,
+	day          = {01},
+	journal      = {Computing},
+	volume       = 71,
+	number       = 1,
+	pages        = {65--87},
+	doi          = {10.1007/s00607-003-0015-5},
+	issn         = {1436-5057},
[…]
