### 2025-09-02 #8300 @baxmittens
@benmkent

Last missing point for me would be

```
Documentation

A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
```

I don't know if there should be a seperate section for this, or if I can consider this fulfilled with the introduction in the documentation.
