### 2025-09-22 #8300 @jbytecode
@benmkent - Our editorialbot fails to parse the bibtex, I think it's because the use of unicode characters in titles and pages. I've just sent a pull request which hopefully solves this. Please review and merge. Ping me when you've done with it. Thank you in advance.

https://github.com/benmkent/SparseGridsKit.jl/pull/35
