### 2025-09-22 #8300 @benmkent
> [@benmkent](https://github.com/benmkent) - It looks like I couldn't address the error. Could you please try to build it locally using docker?
>
> Here is the build script:
>
> #!/bin/bash
>
> docker run --rm \
>     --volume $PWD:/data \
>     --user $(id -u):$(id -g) \
>     --env JOURNAL=joss \
>     openjournals/inara
> Please try to fix the errors, if you can't, I'll ask for a help from the dev team of JOSS. Thank you in advance

yes sure, having a look now.
