+++ paper/paper.bib
@@ -9,7 +9,7 @@
   author = {Bungartz,  Hans-Joachim and Griebel,  Michael},
   year = {2004},
   month = may,
-  pages = {147–269}
+  pages = {147--269}
 }
 
 @book{UQHandbook2017,
@@ -33,7 +33,7 @@
   author = {Schwab,  Christoph and Gittelson,  Claude Jeffrey},
   year = {2011},
   month = apr,
-  pages = {291–467}
+  pages = {291--467}
 }
 
 @article{Piazzola2024,
@@ -48,7 +48,7 @@
   author = {Piazzola,  Chiara and Tamellini,  Lorenzo},
   year = {2024},
   month = mar,
-  pages = {1–22}
+  pages = {1--22}
 }
 
 @article{Cohen2015,
@@ -62,7 +62,7 @@
   author = {Cohen,  Albert and DeVore,  Ronald},
   year = {2015},
   month = apr,
-  pages = {1–159}
+  pages = {1--159}
 }
 
 @book{Sullivan2015,
@@ -217,7 +217,7 @@
   author = {Piazzola,  Chiara and Tamellini,  Lorenzo and Pellegrini,  Riccardo and Broglia,  Riccardo and Serani,  Andrea and Diez,  Matteo},
   year = {2022},
   month = feb,
-  pages = {2209–2237}
+  pages = {2209--2237}
 }
 
 @article{PyApprox,
@@ -252,7 +252,7 @@
   author = {Klimke,  Andreas and Wohlmuth,  Barbara},
   year = {2005},
   month = dec,
-  pages = {561–579}
+  pages = {561--579}
 }
 
 @article{HajiAli2016,
@@ -266,7 +266,7 @@
   author = {Haji-Ali,  Abdul-Lateef and Nobile,  Fabio and Tamellini,  Lorenzo and Tempone,  Raúl},
   year = {2016},
   month = jul,
-  pages = {95–122}
+  pages = {95--122}
 }
 
 @article{Jakeman2019,
@@ -281,7 +281,7 @@
   author = {Jakeman,  John D. and Eldred,  Michael S. and Geraci,  Gianluca and Gorodetsky,  Alex},
   year = {2019},
   month = nov,
-  pages = {1314–1343}
+  pages = {1314--1343}
 }
 
 @inproceedings{Marelli2014,
@@ -293,7 +293,7 @@
   author = {Marelli,  Stefano and Sudret,  Bruno},
   year = {2014},
   month = jun,
-  pages = {2554–2563}
+  pages = {2554--2563}
 }
 
 @TechReport{stoyanov2015tasmanian,
