### 2025-09-22 bc900a5
remove commented bibtex
