+++ paper/paper.bib
@@ -317,18 +317,6 @@
     URL = {http://www5.in.tum.de/pub/pflueger10spatially.pdf}
 }
 
-% @inbook{Debusschere2015,
-%   title = {Uncertainty Quantification Toolkit (UQTk)},
-%   ISBN = {9783319112596},
-%   url = {http://dx.doi.org/10.1007/978-3-319-11259-6_56-1},
-%   DOI = {10.1007/978-3-319-11259-6_56-1},
-%   booktitle = {Handbook of Uncertainty Quantification},
-%   publisher = {Springer International Publishing},
-%   author = {Debusschere,  Bert and Sargsyan,  Khachik and Safta,  Cosmin and Chowdhary,  Kenny},
-%   year = {2015},
-%   pages = {1--21}
-% }
-
 @inbook{DebusschereUQTk:2017,
   title = {Uncertainty Quantification Toolkit (UQTk)},
   ISBN = {9783319123851},
