### 2025-09-22 5abd99a
minor fixes in bibtex
