+++ paper/paper.bib
@@ -104,7 +104,7 @@
 }
 
 @article{Li2024,
-  title = {Characterizing Aquifer Properties through a Sparse Grid-Based Bayesian Framework and InSAR Measurements: A Basin-Scale Application to {A}lto {G}uadalentín,  {S}pain},
+  title = {Characterizing Aquifer Properties through a Sparse Grid-Based {B}ayesian Framework and InSAR Measurements: A Basin-Scale Application to {A}lto {G}uadalentín,  {S}pain},
   url = {http://dx.doi.org/10.22541/essoar.172373105.53381390/v1},
   DOI = {10.22541/essoar.172373105.53381390/v1},
   publisher = {Authorea,  Inc.},
@@ -241,7 +241,7 @@
 }
 
 @article{spinterp,
-  title = {Algorithm 847: {spinterp}: piecewise multilinear hierarchical sparse grid interpolation in {MATLAB}},
+  title = {Algorithm 847: {Spinterp}: piecewise multilinear hierarchical sparse grid interpolation in {MATLAB}},
   volume = {31},
   ISSN = {1557-7295},
   url = {http://dx.doi.org/10.1145/1114268.1114275},
@@ -356,7 +356,7 @@
 }
 
 @article{Bittens2023,
-  title = {DistributedSparseGrids.jl: A Julia library implementing
+  title = {DistributedSparseGrids.jl: A {J}ulia library implementing
 an Adaptive Sparse Grid collocation method},
   volume = {8},
   ISSN = {2475-9066},
