### 2025-09-22 #8300 @jbytecode
@benmkent - The manuscript and the bibliography look better now, thank you.

Now we can go forward. Please perform these actions:

- Please create a tagged release in the software repository, call it e.g. v0.1.0
- Use this tagged release to create a Zenodo archive.
- Report the tagged release, the DOI, and the Zenodo page here.

In Zenodo archive, the archive metadata should match those in the paper (e.g. author list, their ORCIDs, affiliations, the license, etc.)

Please ping me when you've done with them.

Thank you in advance.
