### 2025-09-25 f6a9d47
minor changes in paper.md
