+++ paper/paper.bib
@@ -221,7 +221,7 @@
 }
 
 @article{PyApprox,
- title = {{PyApprox}: A software package for sensitivity analysis, Bayesian inference, optimal experimental design, and multi-fidelity uncertainty quantification and surrogate modeling},
+ title = {{PyApprox}: A software package for sensitivity analysis, {B}ayesian inference, optimal experimental design, and multi-fidelity uncertainty quantification and surrogate modeling},
  journal = {Environmental Modelling \& Software},
  volume = {170},
  pages = {105825},
