### 2025-09-25 #8300 @danielskatz
Congratulations to @benmkent (Benjamin M. Kent) on your JOSS publication!!

And thanks to @dannys4, @baxmittens and @fverdugo for reviewing, and to @jbytecode for editing!
JOSS completely depends on volunteers, and we couldn't do this without you
